[Refactor] 공정 계획 화면 구조 분리
